+++ src/crypto/openssl/verifier.cpp
@@ -56,6 +56,17 @@ namespace crypto
 
     EVP_PKEY* pk = X509_get_pubkey(cert);
 
+#if defined(OPENSSL_VERSION_MAJOR) && OPENSSL_VERSION_MAJOR >= 3
+    auto base_id = EVP_PKEY_get_base_id(pk);
+    if (base_id == EVP_PKEY_EC)
+    {
+      public_key = std::make_unique<PublicKey_OpenSSL>(pk);
+    }
+    else if (base_id == EVP_PKEY_RSA)
+    {
+      public_key = std::make_unique<RSAPublicKey_OpenSSL>(pk);
+    }
+#else
     if (EVP_PKEY_get0_EC_KEY(pk))
     {
       public_key = std::make_unique<PublicKey_OpenSSL>(pk);
@@ -64,6 +75,7 @@ namespace crypto
     {
       public_key = std::make_unique<RSAPublicKey_OpenSSL>(pk);
     }
+#endif
     else
     {
       throw std::logic_error("unsupported public key type");